refactor : 글로벌 익셉션 정의
